git-svn-id: http://svn.ilias.de/svn/ilias/trunk@5778 21b2c9ec-7c21-0410-8b45-9bfb8ed2bfc5

+++ assessment/classes/class.assQuestion.php
@@ -960,7 +960,7 @@ class ASS_Question
 		$deleteresult = $this->ilias->db->query($querydelete);
 				
 		$directory = CLIENT_WEB_DIR . "/assessment/" . $obj_id . "/$question_id";
-		if (is_dir($directory))
+		if (preg_match("/\d+/", $obj_id) and preg_match("/\d+/", $question_id) and is_dir($directory))
 		{
 			$directory = escapeshellarg($directory);
 			exec("rm -rf $directory");
